Move transforms from overlay/ to top level
The transformations used in overlays are generic. To help with using
them from elsewhere, this commit moves them from being a submodule of
overlay/ to the top level, at `@jkcfg/kubernetes/transform`.
